Ecommerce/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==5.1.1
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
